requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
attrs==24.2.0
awscli==1.35.9
botocore==1.35.43
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
docutils==0.16
frozenlist==1.4.1
idna==3.10
jmespath==1.0.1
macholib==1.16.3
multidict==6.1.0
numpy==2.1.2
packaging==24.1
propcache==0.2.0
pyasn1==0.6.1
psycopg2
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.3
setuptools==75.2.0
six==1.16.0
urllib3==2.2.3
yarl==1.15.4
